Move animation logic tests from StoryViewModelTests to StoryAnimationHandlerTests
